Couple of bugfixes for parity with gofmt.

+++ main.go
@@ -148,13 +148,13 @@ func checkPath(path string) (int, error) {
 			resultsJoined := resultsBuf.Bytes()
 
 			funcMid := `) (`
-			funcEnd := `) {`
+			funcEnd := `)`
 			if results == nil || len(results.List) == 0 {
-				funcMid = `) `
-				funcEnd = `{`
+				funcMid = `)`
+				funcEnd = ``
 			} else if len(results.List) == 1 && len(results.List[0].Names) == 0 {
 				funcMid = `) `
-				funcEnd = ` {`
+				funcEnd = ``
 			}
 
 			curFunc.Reset()
@@ -168,14 +168,14 @@ func checkPath(path string) (int, error) {
 				if len(params.List) == 0 {
 					// pass
 				} else if *tab+len(paramsJoined) <= *wrap {
-					fmt.Fprintf(&curFunc, "\n\t%s,", paramsJoined)
+					fmt.Fprintf(&curFunc, "\n\t%s,\n", paramsJoined)
 				} else {
 					for _, param := range params.List {
 						fmt.Fprintf(&curFunc, "\n\t%s,", fileSlice(param.Pos(), param.End()))
 					}
+					curFunc.WriteByte('\n')
 				}
 				if *tab+len(funcMid)+len(resultsJoined)+len(funcEnd) <= *wrap {
-					curFunc.WriteByte('\n')
 					curFunc.WriteString(funcMid)
 					curFunc.Write(resultsJoined)
 					curFunc.WriteString(funcEnd)
@@ -188,6 +188,7 @@ func checkPath(path string) (int, error) {
 					curFunc.WriteString(funcEnd)
 				}
 			}
+			curFunc.Write(fileBytes[fset.Position(f.Type.End()).Offset:fset.Position(closing).Offset])
 
 			oldFunc := fileSlice(opening, closing)
 			if !bytes.Equal(oldFunc, curFunc.Bytes()) {
@@ -209,7 +210,7 @@ func checkPath(path string) (int, error) {
 	}
 	output.Write(fileBytes[fset.Position(lastPos).Offset:])
 
-	if *overwrite {
+	if *overwrite && diffs > 0 {
 		err = ioutil.WriteFile(path, output.Bytes(), 0)
 		if err != nil {
 			return 0, err
